Fix event streams in the orchestrator implementation
