+++ aws/rust-runtime/aws-runtime/Cargo.toml
@@ -7,10 +7,14 @@ edition = "2021"
 license = "Apache-2.0"
 repository = "https://github.com/awslabs/smithy-rs"
 
+[features]
+event-stream = ["dep:aws-smithy-eventstream", "aws-sigv4/sign-eventstream"]
+
 [dependencies]
 aws-credential-types = { path = "../aws-credential-types" }
 aws-http = { path = "../aws-http" }
 aws-sigv4 = { path = "../aws-sigv4" }
+aws-smithy-eventstream = { path = "../../../rust-runtime/aws-smithy-eventstream", optional = true }
 aws-smithy-http = { path = "../../../rust-runtime/aws-smithy-http" }
 aws-smithy-runtime = { path = "../../../rust-runtime/aws-smithy-runtime" }
 aws-smithy-runtime-api = { path = "../../../rust-runtime/aws-smithy-runtime-api" }
@@ -22,6 +26,7 @@ tracing = "0.1"
 uuid = { version = "1", features = ["v4", "fast-rng"] }
 
 [dev-dependencies]
+aws-credential-types = { path = "../aws-credential-types", features = ["test-util"] }
 aws-smithy-protocol-test = { path = "../../../rust-runtime/aws-smithy-protocol-test" }
 proptest = "1"
 serde = { version = "1", features = ["derive"]}
