+++ codegen-core/src/main/kotlin/software/amazon/smithy/rust/codegen/core/smithy/generators/protocol/ProtocolGenerator.kt
@@ -10,6 +10,9 @@ import software.amazon.smithy.rust.codegen.core.rustlang.RustWriter
 import software.amazon.smithy.rust.codegen.core.smithy.CodegenContext
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.Protocol
 
+/** Allows for additional context to be given to the payload generator from where it is being called */
+interface AdditionalPayloadContext
+
 /**
  * Payload Body Generator.
  *
@@ -31,7 +34,10 @@ interface ProtocolPayloadGenerator {
      * Most operations will use the HTTP payload as a reference, but for operations that will consume the entire stream
      * later,they will need to take ownership and different code needs to be generated.
      */
-    fun payloadMetadata(operationShape: OperationShape): PayloadMetadata
+    fun payloadMetadata(
+        operationShape: OperationShape,
+        additionalPayloadContext: AdditionalPayloadContext = object : AdditionalPayloadContext {},
+    ): PayloadMetadata
 
     /**
      * Write the payload into [writer].
@@ -42,7 +48,12 @@ interface ProtocolPayloadGenerator {
      *     - a `Vec<u8>` for non-streaming operations; or
      *     - a `ByteStream` for streaming operations.
      */
-    fun generatePayload(writer: RustWriter, shapeName: String, operationShape: OperationShape)
+    fun generatePayload(
+        writer: RustWriter,
+        shapeName: String,
+        operationShape: OperationShape,
+        additionalPayloadContext: AdditionalPayloadContext = object : AdditionalPayloadContext {},
+    )
 }
 
 /**
